refactor(transformer): move `create_prototype_member` to utils module (#8657)
This function can also used in #8614, so move it.
